// apps/access-server/src/resources/collections.ts
export enum COLLECTION {
  USER = 'users',
  PATIENT = 'patients',
  DOCTOR = 'doctors',
  TEST_CATEGORY = 'test_categories',
  TEST = 'tests',
  TEST_ELEMENT = 'test_elements',
  PATIENT_TYPE = 'patient_types',
  INDICATION = 'indications',
  SAMPLE = 'samples',
  BIO_PRODUCT = 'bio_products',
  SAMPLE_TYPE = 'sample_types',
  TEST_COMBO = 'test_combos',
  PRINT_FORM = 'print_forms',
}